requirements.txt
cryptography==3.0
pyAesCrypt==0.4.3
pycryptodome==3.9.8
